[feat] #13 HomeViewModel 분리
